+++ freppledb/input/views/purchasing.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU Affero General Public
 # License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
-
+from collections import OrderedDict
 
 from django.conf import settings
 from django.db.models.functions import Cast
@@ -385,6 +385,15 @@ class PurchaseOrderList(OperationPlanMixin, GridReport):
 
     @classmethod
     def extra_context(reportclass, request, *args, **kwargs):
+        groupingcfg = OrderedDict()
+        groupingcfg["supplier"] = force_text(_("supplier"))
+        groupingcfg["location"] = force_text(_("location"))
+        groupingcfg["item__category"] = force_text(
+            format_lazy("{} - {}", _("item"), _("category"))
+        )
+        groupingcfg["item__subcategory"] = force_text(
+            format_lazy("{} - {}", _("item"), _("subcategory"))
+        )
         if args and args[0]:
             request.session["lasttab"] = "purchaseorders"
             paths = request.path.split("/")
@@ -397,6 +406,7 @@ class PurchaseOrderList(OperationPlanMixin, GridReport):
                     "model": Supplier,
                     "title": force_text(Supplier._meta.verbose_name) + " " + args[0],
                     "post_title": _("purchase orders"),
+                    "groupingcfg": groupingcfg,
                 }
             elif path == "location" or request.path.startswith(
                 "/detail/input/location/"
@@ -408,6 +418,7 @@ class PurchaseOrderList(OperationPlanMixin, GridReport):
                     "model": Location,
                     "title": force_text(Location._meta.verbose_name) + " " + args[0],
                     "post_title": _("purchase orders"),
+                    "groupingcfg": groupingcfg,
                 }
             elif path == "item" or request.path.startswith("/detail/input/item/"):
                 return {
@@ -417,6 +428,7 @@ class PurchaseOrderList(OperationPlanMixin, GridReport):
                     "model": Item,
                     "title": force_text(Item._meta.verbose_name) + " " + args[0],
                     "post_title": _("purchase orders"),
+                    "groupingcfg": groupingcfg,
                 }
             elif path == "operationplanmaterial":
                 return {
@@ -429,6 +441,7 @@ class PurchaseOrderList(OperationPlanMixin, GridReport):
                         _("on order in %(loc)s at %(date)s")
                         % {"loc": args[1], "date": args[2]}
                     ),
+                    "groupingcfg": groupingcfg,
                 }
             elif path == "produced":
                 return {
@@ -441,6 +454,7 @@ class PurchaseOrderList(OperationPlanMixin, GridReport):
                         _("on order in %(loc)s between %(date1)s and %(date2)s")
                         % {"loc": args[1], "date1": args[2], "date2": args[3]}
                     ),
+                    "groupingcfg": groupingcfg,
                 }
             else:
                 return {
@@ -448,6 +462,7 @@ class PurchaseOrderList(OperationPlanMixin, GridReport):
                     "groupBy": "status",
                     "active_tab": "edit",
                     "model": Item,
+                    "groupingcfg": groupingcfg,
                 }
         elif "parentreference" in request.GET:
             return {
@@ -457,12 +472,14 @@ class PurchaseOrderList(OperationPlanMixin, GridReport):
                 "title": force_text(PurchaseOrder._meta.verbose_name)
                 + " "
                 + request.GET["parentreference"],
+                "groupingcfg": groupingcfg,
             }
         else:
             return {
                 "default_operationplan_type": "PO",
                 "groupBy": "status",
                 "active_tab": "purchaseorders",
+                "groupingcfg": groupingcfg,
             }
 
     @classmethod